Move profile function to porespy.tools.__funcs__
